Remove old icon module
