+++ src/background/mainWindow.js
@@ -2,7 +2,6 @@ import { app, BrowserWindow, ipcMain } from 'electron';
 import createWindowStateKeeper from './windowState';
 import { whenReady, whenReadyToShow } from './utils';
 import env from '../env';
-import icon from './icon';
 
 
 let mainWindow = null;
@@ -83,15 +82,6 @@ export const getMainWindow = async() => {
 		mainWindow.webContents.on('will-navigate', (event) => event.preventDefault());
 		mainWindow.loadURL(`file://${ __dirname }/public/app.html`);
 		attachWindowStateHandling(mainWindow);
-
-		if (process.platform !== 'darwin') {
-			mainWindow.setIcon(await icon.render({
-				size: {
-					win32: [256, 128, 64, 48, 32, 24, 16],
-					linux: 128,
-				}[process.platform],
-			}));
-		}
 
 		if (env.name === 'development') {
 			mainWindow.openDevTools();
