+++ src/scripts/events.js
@@ -177,11 +177,6 @@ export default () => {
 		}
 	});
 
-	webview.on('ipc-message-user-status-manually-set', (hostUrl, [status]) => {
-		tray.setState({ status });
-		dock.setState({ status });
-	});
-
 	if (process.platform === 'darwin') {
 		setTouchBar();
 	}
